moved cluster apis
