+++ api/agent/handlers.go
@@ -888,152 +888,3 @@ func (s *Server) tornjakAgentsList(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 }
-
-/********* CLUSTER *********/
-
-func (s *Server) clusterList(w http.ResponseWriter, r *http.Request) {
-	var input ListClustersRequest
-	buf := new(strings.Builder)
-
-	n, err := io.Copy(buf, r.Body)
-	if err != nil {
-		emsg := fmt.Sprintf("Error parsing data: %v", err.Error())
-		retError(w, emsg, http.StatusBadRequest)
-		return
-	}
-	data := buf.String()
-
-	if n == 0 {
-		input = ListClustersRequest{}
-	} else {
-		err := json.Unmarshal([]byte(data), &input)
-		if err != nil {
-			emsg := fmt.Sprintf("Error parsing data: %v", err.Error())
-			retError(w, emsg, http.StatusBadRequest)
-			return
-		}
-	}
-
-	ret, err := s.ListClusters(input)
-	if err != nil {
-		emsg := fmt.Sprintf("Error: %v", err.Error())
-		retError(w, emsg, http.StatusBadRequest)
-		return
-	}
-	cors(w, r)
-	je := json.NewEncoder(w)
-	err = je.Encode(ret)
-	if err != nil {
-		emsg := fmt.Sprintf("Error: %v", err.Error())
-		retError(w, emsg, http.StatusBadRequest)
-		return
-	}
-}
-
-func (s *Server) clusterCreate(w http.ResponseWriter, r *http.Request) {
-	buf := new(strings.Builder)
-	n, err := io.Copy(buf, r.Body)
-	if err != nil {
-		emsg := fmt.Sprintf("Error parsing data: %v", err.Error())
-		retError(w, emsg, http.StatusBadRequest)
-		return
-	}
-	data := buf.String()
-	var input RegisterClusterRequest
-	if n == 0 {
-		input = RegisterClusterRequest{}
-	} else {
-		err := json.Unmarshal([]byte(data), &input)
-		if err != nil {
-			emsg := fmt.Sprintf("Error parsing data: %v", err.Error())
-			retError(w, emsg, http.StatusBadRequest)
-			return
-		}
-	}
-	err = s.DefineCluster(input)
-	if err != nil {
-		emsg := fmt.Sprintf("Error: %v", err.Error())
-		retError(w, emsg, http.StatusBadRequest)
-		return
-	}
-	cors(w, r)
-	_, err = w.Write([]byte("SUCCESS"))
-	if err != nil {
-		emsg := fmt.Sprintf("Error: %v", err.Error())
-		retError(w, emsg, http.StatusBadRequest)
-		return
-	}
-}
-
-func (s *Server) clusterEdit(w http.ResponseWriter, r *http.Request) {
-	buf := new(strings.Builder)
-	n, err := io.Copy(buf, r.Body)
-	if err != nil {
-		emsg := fmt.Sprintf("Error parsing data: %v", err.Error())
-		retError(w, emsg, http.StatusBadRequest)
-		return
-	}
-	data := buf.String()
-	var input EditClusterRequest
-	if n == 0 {
-		input = EditClusterRequest{}
-	} else {
-		err := json.Unmarshal([]byte(data), &input)
-		if err != nil {
-			emsg := fmt.Sprintf("Error parsing data: %v", err.Error())
-			retError(w, emsg, http.StatusBadRequest)
-			return
-		}
-	}
-	err = s.EditCluster(input)
-	if err != nil {
-		emsg := fmt.Sprintf("Error: %v", err.Error())
-		retError(w, emsg, http.StatusBadRequest)
-		return
-	}
-	cors(w, r)
-	_, err = w.Write([]byte("SUCCESS"))
-	if err != nil {
-		emsg := fmt.Sprintf("Error: %v", err.Error())
-		retError(w, emsg, http.StatusBadRequest)
-		return
-	}
-}
-
-func (s *Server) clusterDelete(w http.ResponseWriter, r *http.Request) {
-	buf := new(strings.Builder)
-	n, err := io.Copy(buf, r.Body)
-	if err != nil {
-		emsg := fmt.Sprintf("Error parsing data: %v", err.Error())
-		retError(w, emsg, http.StatusBadRequest)
-		return
-	}
-	data := buf.String()
-	var input DeleteClusterRequest
-	if n == 0 {
-		input = DeleteClusterRequest{}
-	} else {
-		err := json.Unmarshal([]byte(data), &input)
-		if err != nil {
-			emsg := fmt.Sprintf("Error parsing data: %v", err.Error())
-			retError(w, emsg, http.StatusBadRequest)
-			return
-		}
-	}
-	err = s.DeleteCluster(input)
-	if err != nil {
-		emsg := fmt.Sprintf("Error: %v", err.Error())
-		retError(w, emsg, http.StatusBadRequest)
-		return
-	}
-	cors(w, r)
-	_, err = w.Write([]byte("SUCCESS"))
-	if err != nil {
-		emsg := fmt.Sprintf("Error: %v", err.Error())
-		retError(w, emsg, http.StatusBadRequest)
-		return
-	}
-
-}
-
-/********* END CLUSTER *********/
